+++ src/constants.py
@@ -1,8 +1,8 @@
 class Null:
 	__slots__ = ()
 
-AI = 0x1000
-ML = 0x1001
-
 TG_INPUT = "inp"  # tag input
 TG_OUTPUT = "out"  # tag output
+FLD_CONST = "const"  # const field
+
+DEBUG__ = True